requirements.txt
Django==3.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
boto3==1.20.45
sentry-sdk==1.5.3
python-dotenv==0.19.2
PyJWT==2.3.0
stringcase==1.2.0